refactor: remove remaining U1000 dead code across adapter/daemon/install
- Delete 67 unused production funcs/consts/types flagged by staticcheck U1000
  (adapter/claude+codex+feishu, app/daemon, app/install, core/control,
  claudesessionstore, config, issuedocsync, issueworkflow)
- Remove 33 unused test-only helper vars/functions from _test.go files
- Relocate 12 card/git helpers referenced only by tests into a shared
  _test.go file (card_transport_helpers_test.go) so production code stays
  free of test-only surface
- staticcheck -checks U1000 ./... now reports 0 findings; go build and
  go vet pass; full go test ./... has no new failures (two pre-existing
  environment failures: Windows symlink privilege in orchestrator,
  resume-notice timing on master)

Refs: #818, #807
