+++ internal/adapter/feishu/gateway_helpers.go
@@ -75,77 +75,3 @@ func mimeExtension(mimeType string) string {
 		return ""
 	}
 }
-
-func stringMapValue(values map[string]interface{}, key string) string {
-	if len(values) == 0 {
-		return ""
-	}
-	value, ok := values[key]
-	if !ok || value == nil {
-		return ""
-	}
-	switch current := value.(type) {
-	case string:
-		return current
-	default:
-		return fmt.Sprint(current)
-	}
-}
-
-func boolMapValue(values map[string]interface{}, key string) bool {
-	if len(values) == 0 {
-		return false
-	}
-	value, ok := values[key]
-	if !ok || value == nil {
-		return false
-	}
-	current, _ := value.(bool)
-	return current
-}
-
-func intMapValue(values map[string]interface{}, key string) int {
-	if len(values) == 0 {
-		return 0
-	}
-	value, ok := values[key]
-	if !ok || value == nil {
-		return 0
-	}
-	switch current := value.(type) {
-	case int:
-		return current
-	case int8:
-		return int(current)
-	case int16:
-		return int(current)
-	case int32:
-		return int(current)
-	case int64:
-		return int(current)
-	case uint:
-		return int(current)
-	case uint8:
-		return int(current)
-	case uint16:
-		return int(current)
-	case uint32:
-		return int(current)
-	case uint64:
-		return int(current)
-	case float32:
-		return int(current)
-	case float64:
-		return int(current)
-	case string:
-		current = strings.TrimSpace(current)
-		if current == "" {
-			return 0
-		}
-		var parsed int
-		_, _ = fmt.Sscanf(current, "%d", &parsed)
-		return parsed
-	default:
-		return 0
-	}
-}
