+++ internal/adapter/feishu/preview/markdown_preview_managed.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"log"
 	"sort"
-	"strconv"
 	"strings"
 	"time"
 )
@@ -37,27 +36,6 @@ func previewRemoteCleanupTime(node previewRemoteNode) (time.Time, bool) {
 	default:
 		return time.Time{}, false
 	}
-}
-
-func parsePreviewRemoteTime(raw string) time.Time {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		return time.Time{}
-	}
-	if value, err := strconv.ParseInt(raw, 10, 64); err == nil {
-		switch {
-		case value > 1_000_000_000_000:
-			return time.UnixMilli(value).UTC()
-		case value > 0:
-			return time.Unix(value, 0).UTC()
-		default:
-			return time.Time{}
-		}
-	}
-	if value, err := time.Parse(time.RFC3339, raw); err == nil {
-		return value.UTC()
-	}
-	return time.Time{}
 }
 
 func (p *DriveMarkdownPreviewer) cleanupManagedPreviewFiles(ctx context.Context, cutoff time.Time) (PreviewDriveCleanupResult, error) {
